+++ GermanVocabApp.Api/VocabLists/Validation/VocabListItems/CreateItemValidator.cs
@@ -1,6 +1,6 @@
 ﻿using GermanVocabApp.Api.VocabLists.Models;
 
-namespace GermanVocabApp.Api.VocabLists.Validation;
+namespace GermanVocabApp.Api.VocabLists.Validation.ListItems;
 
 public class CreateItemValidator : AbstractItemValidator<CreateVocabListItemRequest>
 {
