+++ GermanVocabApp.Api/VocabLists/Validation/VocabLists/CreateListValidator.cs
@@ -1,7 +1,7 @@
 ﻿using GermanVocabApp.Api.VocabLists.Models;
 using GermanVocabApp.Core.Validation.DependencyInjection;
 
-namespace GermanVocabApp.Api.VocabLists.Validation;
+namespace GermanVocabApp.Api.VocabLists.Validation.VocabLists;
 
 public class CreateListValidator : AbstractListValidator<CreateVocabListRequest, CreateVocabListItemRequest>
 {
