+++ GermanVocabApp.Api/VocabLists/Validation/VocabLists/UpdateListValidator.cs
@@ -1,7 +1,7 @@
 ﻿using GermanVocabApp.Api.VocabLists.Models;
 using GermanVocabApp.Core.Validation.DependencyInjection;
 
-namespace GermanVocabApp.Api.VocabLists.Validation;
+namespace GermanVocabApp.Api.VocabLists.Validation.VocabLists;
 
 public class UpdateListValidator : AbstractListValidator<UpdateVocabListRequest, UpdateVocabListItemRequest>
 {
